+++ Pipfile
@@ -26,6 +26,7 @@ pandas = "*"
 tqdm = "*"
 ray = "*"
 flake8 = "*"
+ssl-bench = {file = ".", editable = true}
 
 [torch_cpu]
 torch = {index="pytorch_cpu"}
