+++ pyproject.toml
@@ -1,0 +1,3 @@
+[project]
+name = "ssl_bench"
+version = "0.0.1"
