+++ src/ssl_bench/LabelOnlyBaseline/LabelOnlyBaseline.py
@@ -28,12 +28,11 @@ from torchvision import transforms
 
 from torch.utils.tensorboard import SummaryWriter
 
-from src.dataset import data as dataset
-from src.randaugment import RandAugmentMC
-from src.CoMatch.libml.utils import save_pickle
-from src.CoMatch.libml.utils import train_one_epoch, eval_model
-from src.CoMatch.libml.utils import EarlyStopping
-from src.config import dataset_configs
+from ssl_bench.dataset import data as dataset
+from ssl_bench.LabelOnlyBaseline.libml.utils import save_pickle
+from ssl_bench.LabelOnlyBaseline.libml.utils import train_one_epoch, eval_model
+from ssl_bench.LabelOnlyBaseline.libml.utils import EarlyStopping
+from ssl_bench.config import dataset_configs
 
 
 
@@ -42,7 +41,7 @@ logger = logging.getLogger(__name__)
 
 parser = argparse.ArgumentParser()
 #experiment setting
-parser.add_argument('--dataset_name', default='TissueMNIST', type=str, help='name of dataset')
+parser.add_argument('--dataset_name', default='PathMNIST', type=str, help='name of dataset')
 parser.add_argument('--data_seed', default=0, type=int, help='random seed data partitioning procedure')
 parser.add_argument('--training_seed', default=0, type=int, help='random seed for training procedure')
 
@@ -64,16 +63,13 @@ parser.add_argument('--train_dir',
 
 #data paths
 parser.add_argument('--l_train_dataset_path', default='', type=str)
-parser.add_argument('--u_train_dataset_path', default='', type=str)
 parser.add_argument('--val_dataset_path', default='', type=str)
 parser.add_argument('--test_dataset_path', default='', type=str)
 
 #shared config
 parser.add_argument('--labeledtrain_batchsize', default=50, type=int)
-parser.add_argument('--unlabeledtrain_batchsize', default=50, type=int)
-parser.add_argument("--em", default=0, type=float, help="coefficient of entropy minimization. If you try VAT + EM, set 0.06")
 
-#FM config
+#VAT config
 parser.add_argument('--lr', default=3e-4, type=float, help='learning rate')
 parser.add_argument('--lr_warmup_epochs', default=0, type=float,
                     help='warmup epoch for learning rate schedule') #following MixMatch and FixMatch repo
@@ -81,30 +77,18 @@ parser.add_argument('--lr_warmup_epochs', default=0, type=float,
 parser.add_argument('--lr_schedule_type', default='CosineLR', choices=['CosineLR', 'FixedLR'], type=str) 
 parser.add_argument('--lr_cycle_epochs', default=10000, type=int) #following MixMatch and FixMatch repo
 
+
 parser.add_argument('--wd', default=5e-4, type=float, help='weight decay')
 parser.add_argument('--optimizer_type', default='SGD', choices=['SGD', 'Adam'], type=str) 
-
-
-parser.add_argument('--temperature', default=0.2, type=float, help='temperature for label guessing')#note COMATCH set a verly low temperature
-
-parser.add_argument('--mu', default=7, type=int,
-                    help='coefficient of unlabeled batch size')
-
-parser.add_argument('--threshold', default=0.95, type=float,
-                    help='pseudo label threshold')
-
-
-parser.add_argument('--lambda_u_max', default=1, type=float, help='coefficient of unlabeled loss')
-
-parser.add_argument('--unlabeledloss_warmup_schedule_type', default='NoWarmup', choices=['NoWarmup', 'Linear', 'Sigmoid', ], type=str) 
-
-parser.add_argument('--unlabeledloss_warmup_pos', default=0.4, type=float, help='position at which unlabeled loss warmup ends') #following MixMatch and FixMatch repo
 
 
 
 #default hypers not to search for now
 parser.add_argument('--nesterov', action='store_true', default=True,
                     help='use nesterov momentum')
+
+parser.add_argument('--use_ema', action='store_true', default=True,
+                    help='use EMA model')
 
 parser.add_argument('--ema_decay', default=0.999, type=float,
                     help='EMA decay rate')
@@ -118,17 +102,6 @@ parser.add_argument('--use_pretrained', default='False', type=str)
 
 #early stopping
 parser.add_argument('--patience', default=20, type=int, help='Earlystop patience')
-
-
-#####Unique to COMATCH than FixMathc!!!!!!!!!!!!
-parser.add_argument('--alpha', type=float, default=0.9)  
-parser.add_argument('--queue-batch', type=float, default=5, 
-                        help='number of batches stored in memory bank')
-parser.add_argument('--contrast-th', default=0.8, type=float,
-                        help='pseudo label graph threshold')   
-parser.add_argument('--lam-c', type=float, default=1,
-                        help='coefficient of contrastive loss')  
-parser.add_argument('--low-dim', type=int, default=64)
 
 
 def sample_loguniform(low=0, high=1, size=1, coefficient=1, base=10):
@@ -192,14 +165,16 @@ def get_fixed_lr(optimizer,
     return LambdaLR(optimizer, _lr_lambda, last_epoch)    
 
 
-
 def create_model(args):
-    if args.arch=='resnet18':
-        import backbone.resnet18_comatch as models
-        model = models.build_resnet18(args.num_classes, pretrained=args.use_pretrained, progress=True)
-    
+    if args.arch == 'resnet18':
+        from torchvision import models
+        
+        model = models.resnet18(pretrained=args.use_pretrained)
+        #https://stackoverflow.com/questions/52548174/how-to-remove-the-last-fc-layer-from-a-resnet-model-in-pytorch
+        model.fc = torch.nn.Linear(512, args.num_classes)
+        
     elif args.arch=='wideresnet':
-        import backbone.wideresnet_comatch as models
+        import backbone.wideresnet as models
         model_depth = 28
         model_width = 2
             
@@ -207,19 +182,22 @@ def create_model(args):
                                         widen_factor=model_width,
                                         dropout=0.0,
                                         num_classes=args.num_classes)
-            
+        
     else:
         raise NameError('Note implemented yet')
-        
+    
+
+    
     logger.info("Total params: {:.2f}M".format(
         sum(p.numel() for p in model.parameters())/1e6))
     return model
 
+    
 
 
 
 def main(args):
-    
+            
     precalculated_class_weights = dataset_configs[args.dataset_name]['class_weights']
     weights = torch.Tensor(precalculated_class_weights)
 #     print('weights used is {}'.format(weights))
@@ -228,6 +206,7 @@ def main(args):
     dataset_mean = dataset_configs[args.dataset_name]['dataset_mean']
     dataset_std = dataset_configs[args.dataset_name]['dataset_std']
     image_size = dataset_configs[args.dataset_name]['image_size']
+    
     
     transform_labeledtrain = transforms.Compose([
         transforms.RandomHorizontalFlip(),
@@ -243,130 +222,72 @@ def main(args):
         transforms.Normalize(mean=dataset_mean, std=dataset_std)
     ])
     
-    
-    class TransformCoMatch(object):
-        def __init__(self, mean, std):
-            self.weak = transforms.Compose([
-                transforms.RandomHorizontalFlip(),
-                transforms.RandomCrop(size=image_size,
-                                      padding=int(image_size*0.125),
-                                      padding_mode='reflect')])
-            
-            self.strong0 = transforms.Compose([
-                transforms.RandomHorizontalFlip(),
-                transforms.RandomCrop(size=image_size,
-                                      padding=int(image_size*0.125),
-                                      padding_mode='reflect'),
-                RandAugmentMC(n=2, m=10)])
-            
-            self.strong1 = transforms.Compose([
-                transforms.RandomHorizontalFlip(),
-                transforms.RandomCrop(size=image_size,
-                                      padding=int(image_size*0.125),
-                                      padding_mode='reflect'),
-                transforms.RandomApply([
-                transforms.ColorJitter(0.4, 0.4, 0.4, 0.1)  
-                ], p=0.8),
-                transforms.RandomGrayscale(p=0.2)])
-            
-            self.normalize = transforms.Compose([
-                transforms.ToTensor(),
-                transforms.Normalize(mean=mean, std=std)])
+  
 
-        def __call__(self, x):
-            weak = self.weak(x)
-            strong0 = self.strong0(x)
-            strong1 = self.strong1(x)
-            return self.normalize(weak), self.normalize(strong0), self.normalize(strong1)
-
-      
     l_train_dataset = dataset(args.dataset_name, args.l_train_dataset_path, transform_fn=transform_labeledtrain)
-    u_train_dataset = dataset(args.dataset_name, args.u_train_dataset_path, transform_fn=TransformCoMatch(mean=dataset_mean, std=dataset_std))
     val_dataset = dataset(args.dataset_name, args.val_dataset_path, transform_fn=transform_eval)
     test_dataset = dataset(args.dataset_name, args.test_dataset_path, transform_fn=transform_eval)
-        
-        
-    #########################################for loop from here##################################      
+    
+    #########################################for loop from here##################################
+    
     global_best_val_raw_acc_list_parallel = []
     global_best_test_raw_acc_at_val_list_parallel = []
-    
+
     
     global_best_val_raw_acc = 0
-    global_best_test_raw_acc_at_val = 0    
+    global_best_test_raw_acc_at_val = 0
+#     global_best_train_raw_acc_at_val = 0
 
+    
     #create global writer
     global_writer = SummaryWriter(args.train_dir)
 
-   
+    
     #start timing
     hypercombo_iteratethrough_list = [] #newly added
     hypercombo_iteratethrough_time_list = []
-    
-    
+
     start_time = time.time()
     total_used_time = 0
     
     #Lr: loguniform base 10, [3*10^(-5) to 3*10^(-2)]
     #Wd: loguniform base 10, [4*10^(-6) to 4*10^(-3)]
-    #Lambda_u_max: loguniform base 10, [1*10^(-1) to 1*10^(1)]
-    
+    #Lambda_u_max: loguniform base 10, [8*10^(-1) to 8*10^(1)]
     while total_used_time <= args.total_hour * 3600: #run for 100 hour
-        
-        #initialize for every model
-        queue_feats = torch.zeros(args.queue_size, args.low_dim).cuda()
-        queue_probs = torch.zeros(args.queue_size, args.num_classes).cuda()
-        queue_ptr = 0
-    
-        # for distribution alignment
-        prob_list = []
-    
         lr = sample_loguniform(low=-5, high=-2, size=1, coefficient=3, base=10)
         wd = sample_loguniform(low=-6, high=-3, size=1, coefficient=4, base=10)
-        lambda_u_max = sample_loguniform(low=-1, high=1, size=1, coefficient=1, base=10)
-        lam_c = sample_loguniform(low=-1, high=1, size=1, coefficient=5, base=10)
         
+        print('!!!!!!!!!!This run using lr: {}, wd: {}!!!!!!!!!!'.format(lr, wd))
         
-        print('!!!!!!!!!!This run using lr: {}, wd: {}, lambda_u_max: {}, lam_c: {}!!!!!!!!!!'.format(lr, wd, lambda_u_max, lam_c))
-        hypercombo_iteratethrough_list.append({'lr':lr, 'wd':wd, 'lambda_u_max':lambda_u_max, 'lam_c':lam_c})#newly added
+        hypercombo_iteratethrough_list.append({'lr':lr, 'wd':wd})#newly added
         save_pickle(os.path.join(args.train_dir, 'global_stats'), 'hypercombo_iteratethrough_list.pkl', hypercombo_iteratethrough_list)
         this_hypercombo_starttime = time.time()
         
         this_hypercombo_best_val_raw_acc_list_parallel = []
         this_hypercombo_best_test_raw_acc_at_val_list_parallel = []
+
         
         args.lr = lr
         args.wd = wd
-        args.lambda_u_max = lambda_u_max
-        args.lam_c = lam_c
-        
 
-
-        experiment_name = "LambdaUMax-{}-LamC-{}_lr-{}_wd-{}_ThresholdC-{}_alpha-{}_temperature-{}_mu-{}_threshold-{}".format(args.lambda_u_max, args.lam_c, args.lr, args.wd, args.contrast_th, args.alpha, args.temperature, args.mu, args.threshold)
+        experiment_name = "lr-{}_wd-{}".format(args.lr, args.wd)
 
         args.experiment_dir = os.path.join(args.train_dir, 'hypercombos', experiment_name)
 
         #brief summary:
         brief_summary = {}
         brief_summary['dataset_name'] = args.dataset_name
-        brief_summary['algorithm'] = 'CoMatch'
+        brief_summary['algorithm'] = 'Supervised'
         brief_summary['hyperparameters'] = {
             'optimizer': args.optimizer_type,
             'lr_schedule_type': args.lr_schedule_type,
             'lr_cycle_epochs': args.lr_cycle_epochs,
-            'unlabeledloss_warmup_schedule_type':args.unlabeledloss_warmup_schedule_type,
-            'unlabeledloss_warmup_pos': args.unlabeledloss_warmup_pos,
-            'lambda_u_max': args.lambda_u_max,  
-            'lam_c':args.lam_c,
             'lr': args.lr,
             'wd': args.wd,
-            'contrast_th':args.contrast_th,
-            'alpha':args.alpha,
-            'temperature': args.temperature,
-            'mu': args.mu,
-            'threshold':args.threshold,
+
         }
 
-
+        
         if args.resume != 'None':
             args.resume_checkpoint_fullpath = os.path.join(args.experiment_dir, args.resume)
             print('args.resume_checkpoint_fullpath: {}'.format(args.resume_checkpoint_fullpath))
@@ -377,27 +298,28 @@ def main(args):
         args.writer = SummaryWriter(args.experiment_dir)
 
         print('Created dataset')
-        print("labeled data : {}, unlabeled data : {}, training data : {}".format(
-        len(l_train_dataset), len(u_train_dataset), len(l_train_dataset)+len(u_train_dataset)))
-
+        print("labeled data : {}, training data : {}".format(
+        len(l_train_dataset), len(l_train_dataset)))
+        print("validation data : {}, test data : {}".format(len(val_dataset), len(test_dataset)))
+    
 
         l_loader = DataLoader(l_train_dataset, args.labeledtrain_batchsize, shuffle=True, drop_last=True, num_workers=args.num_workers)
-        u_loader = DataLoader(u_train_dataset, args.unlabeledtrain_batchsize, shuffle=True, drop_last=True, num_workers=args.num_workers)
         val_loader = DataLoader(val_dataset, 128, shuffle=False, drop_last=False, num_workers=args.num_workers)
         test_loader = DataLoader(test_dataset, 128, shuffle=False, drop_last=False, num_workers=args.num_workers)
 
         #create model
-        model = create_model(args) 
+        model = create_model(args)
         model.to(args.device)
+
 
         #optimizer_type choice
         no_decay = ['bias', 'bn']
         grouped_parameters = [
-                {'params': [p for n, p in model.named_parameters() if not any(
-                    nd in n for nd in no_decay)], 'weight_decay': args.wd},
-                {'params': [p for n, p in model.named_parameters() if any(
-                    nd in n for nd in no_decay)], 'weight_decay': 0.0}
-            ]
+            {'params': [p for n, p in model.named_parameters() if not any(
+                nd in n for nd in no_decay)], 'weight_decay': args.wd},
+            {'params': [p for n, p in model.named_parameters() if any(
+                nd in n for nd in no_decay)], 'weight_decay': 0.0}
+        ]
 
         if args.optimizer_type == 'SGD':
             optimizer = optim.SGD(grouped_parameters, lr=args.lr,
@@ -405,9 +327,9 @@ def main(args):
 
         elif args.optimizer_type == 'Adam':
             optimizer = optim.Adam(grouped_parameters, lr=args.lr)
+
         else:
             raise NameError('Not supported optimizer setting')
-
 
 
         #lr_schedule_type choice
@@ -421,7 +343,7 @@ def main(args):
             raise NameError('Not supported lr scheduler setting')
 
 
-#         #instantiate the ema model object
+        #instantiate the ema model object
 #         ema_model = ModelEMA(args, model, args.ema_decay)
 
         args.start_epoch = 0
@@ -433,8 +355,10 @@ def main(args):
         this_hypercombo_best_val_raw_acc_list_parallel.append(best_val_raw_acc)
         this_hypercombo_best_test_raw_acc_at_val_list_parallel.append(best_test_raw_acc_at_val)
 
-        current_count=0 #for early stopping, when continue training   
         
+        current_count=0 #for early stopping, when continue training
+
+
         if args.resume_checkpoint_fullpath is not None:
             try:
                 os.path.isfile(args.resume_checkpoint_fullpath)
@@ -442,11 +366,10 @@ def main(args):
                 checkpoint = torch.load(args.resume_checkpoint_fullpath)
                 args.start_epoch = checkpoint['epoch']
                 model.load_state_dict(checkpoint['state_dict'])
-#                 ema_model.ema.load_state_dict(checkpoint['ema_state_dict'])
 
                 best_val_raw_acc = checkpoint['best_val_raw_acc']
                 best_test_raw_acc_at_val = checkpoint['best_test_raw_acc_at_val']
-#                         best_train_raw_acc_at_val = checkpoint['best_train_raw_acc_at_val']
+    #             best_train_raw_acc_at_val = checkpoint['best_train_raw_acc_at_val']
 
                 current_count = checkpoint['current_count']
                 optimizer.load_state_dict(checkpoint['optimizer'])
@@ -458,7 +381,7 @@ def main(args):
         logger.info("***** Running training *****")
         logger.info(f"  Task = {args.dataset_name}")
         logger.info(f"  Num Epochs = {args.train_epoch}")
-        logger.info(f"  Batch size per GPU (labeled+unlabeled) = {args.labeledtrain_batchsize + args.unlabeledtrain_batchsize}")
+        logger.info(f"  Batch size per GPU (labeled) = {args.labeledtrain_batchsize}")
 
 
         is_best = False
@@ -467,24 +390,25 @@ def main(args):
 
 
         for epoch in range(args.start_epoch, args.train_epoch):
-
+            
             #train
-            train_total_loss_list, train_labeled_loss_list, train_unlabeled_loss_unscaled_list, train_unlabeled_loss_scaled_list, train_contrastive_loss_list, queue_feats, queue_probs, queue_ptr, prob_list = train_one_epoch(args, weights, l_loader, u_loader, model, optimizer, scheduler, epoch, prob_list, queue_feats, queue_probs, queue_ptr)
+            train_labeled_loss_list = train_one_epoch(args, weights, l_loader, model, optimizer, scheduler, epoch)
 
-
+            
             #val
             val_loss, val_raw_acc, val_true_labels, val_raw_predictions = eval_model(args, val_loader, model, epoch, evaluation_criterion='balanced_accuracy')
 
             #test
             test_loss, test_raw_acc, test_true_labels, test_raw_predictions = eval_model(args, test_loader, model, epoch, evaluation_criterion='balanced_accuracy')
 
-
             if val_raw_acc > best_val_raw_acc:
-                is_best=True
+                is_best = True
 
                 best_val_raw_acc = val_raw_acc
                 best_test_raw_acc_at_val = test_raw_acc
 #                 best_train_raw_acc_at_val = train_raw_acc
+
+
 
 
             this_hypercombo_best_val_raw_acc_list_parallel.append(best_val_raw_acc)
@@ -496,14 +420,9 @@ def main(args):
 
             start_time = time.time() #reinitialize
 
-            logger.info('RAW Best , validation/test %.2f %.2f' % (best_val_raw_acc, best_test_raw_acc_at_val))
+            logger.info('RAW Best , validation/test/train %.2f %.2f' % (best_val_raw_acc, best_test_raw_acc_at_val))
 
-            args.writer.add_scalar('train/3.total_loss', np.mean(train_total_loss_list), epoch)
             args.writer.add_scalar('train/4.labeled_loss', np.mean(train_labeled_loss_list), epoch)
-            args.writer.add_scalar('train/5.unlabeled_loss_unscaled', np.mean(train_unlabeled_loss_unscaled_list), epoch)
-            args.writer.add_scalar('train/6.unlabele_loss_scaled', np.mean(train_unlabeled_loss_scaled_list), epoch)
-            args.writer.add_scalar('train/7.contrastive_loss', np.mean(train_contrastive_loss_list), epoch)
-
 
             args.writer.add_scalar('val/1.val_raw_acc', val_raw_acc, epoch)
             args.writer.add_scalar('val/3.val_loss', val_loss, epoch)
@@ -511,14 +430,16 @@ def main(args):
             args.writer.add_scalar('test/3.test_loss', test_loss, epoch)
 
             brief_summary["number_of_data"] = {
-        "labeled":len(l_train_dataset), "unlabeled":len(u_train_dataset),
+        "labeled":len(l_train_dataset), 
         "validation":len(val_dataset), "test":len(test_dataset)
     }
+
             brief_summary['best_test_raw_acc_at_val'] = best_test_raw_acc_at_val
             brief_summary['best_val_raw_acc'] = best_val_raw_acc
 
             with open(os.path.join(args.experiment_dir + "brief_summary.json"), "w") as f:
                 json.dump(brief_summary, f)
+
 
             #early stopping
             current_count = early_stopping(val_raw_acc)
@@ -543,18 +464,19 @@ def main(args):
 
             if total_used_time > args.total_hour * 3600:
                 break
-                    
 
 
         this_hypercombo_endtime = time.time()
         this_hypercombo_iteratethrough_time = this_hypercombo_endtime - this_hypercombo_starttime
         hypercombo_iteratethrough_time_list.append(this_hypercombo_iteratethrough_time)#newly added
         save_pickle(os.path.join(args.train_dir, 'global_stats'), 'hypercombo_iteratethrough_time_list.pkl', hypercombo_iteratethrough_time_list)
+
         
         brief_summary["number_of_data"] = {
-        "labeled":len(l_train_dataset), "unlabeled":len(u_train_dataset),
+        "labeled":len(l_train_dataset), 
         "validation":len(val_dataset), "test":len(test_dataset)
     }
+        
         brief_summary['best_test_raw_acc_at_val'] = best_test_raw_acc_at_val
         brief_summary['best_val_raw_acc'] = best_val_raw_acc
         
@@ -562,27 +484,28 @@ def main(args):
         global_best_val_raw_acc_list_parallel.append(this_hypercombo_best_val_raw_acc_list_parallel)
         global_best_test_raw_acc_at_val_list_parallel.append(this_hypercombo_best_test_raw_acc_at_val_list_parallel)
 
-
+        
         save_pickle(os.path.join(args.train_dir, 'global_stats'), 'global_best_val_raw_acc_list_parallel.pkl', global_best_val_raw_acc_list_parallel)
         save_pickle(os.path.join(args.train_dir, 'global_stats'), 'global_best_test_raw_acc_at_val_list_parallel.pkl', global_best_test_raw_acc_at_val_list_parallel)
+        
 
         args.writer.close()
 
         with open(os.path.join(args.experiment_dir + "brief_summary.json"), "w") as f:
             json.dump(brief_summary, f)
-            
-    save_pickle(os.path.join(args.train_dir, 'global_stats'), 'total_time.pkl', [total_used_time])
 
             
+    save_pickle(os.path.join(args.train_dir, 'global_stats'), 'total_time.pkl', [total_used_time])
+    
+    
+                    
 if __name__ == '__main__':
     args = parser.parse_args()
     
     args.use_pretrained = str2bool(args.use_pretrained)
     args.nimg_per_epoch = dataset_configs[args.dataset_name]['nimg_per_epoch'] #total size of labeled + unlabeled set for TissueMNIST
     args.num_classes = dataset_configs[args.dataset_name]['num_classes']
-   
-    # memory bank
-    args.queue_size = args.queue_batch*(args.mu+1)*args.labeledtrain_batchsize
+    
     
     cuda = torch.cuda.is_available()
     
@@ -603,15 +526,9 @@ if __name__ == '__main__':
     logger.info(dict(args._get_kwargs()))
 
     
+    
     if args.training_seed is not None:
         print('setting training seed{}'.format(args.training_seed), flush=True)
         set_seed(args.training_seed)
-
+        
     main(args)
-    
-    
-    
-    
-
-
-
