+++ src/ssl_bench/MeanTeacher/MeanTeacher.py
@@ -28,12 +28,15 @@ from torchvision import transforms
 
 from torch.utils.tensorboard import SummaryWriter
 
-from src.dataset import data as dataset
-from src.LabelOnlyBaseline.libml.utils import save_pickle
-from src.LabelOnlyBaseline.libml.utils import train_one_epoch, eval_model
-from src.LabelOnlyBaseline.libml.utils import EarlyStopping
-from src.config import dataset_configs
+from ssl_bench.dataset import data as dataset
+from ssl_bench.MeanTeacher.libml.utils import save_pickle
+from ssl_bench.MeanTeacher.libml.utils import train_one_epoch, eval_model
+from ssl_bench.MeanTeacher.libml.utils import EarlyStopping
+from ssl_bench.config import dataset_configs
+from ssl_bench.MeanTeacher.libml.utils.ema import ModelEMA
 
+
+from ssl_bench.MeanTeacher.libml.utils.mean_teacher import MT
 
 
 logger = logging.getLogger(__name__)
@@ -41,7 +44,7 @@ logger = logging.getLogger(__name__)
 
 parser = argparse.ArgumentParser()
 #experiment setting
-parser.add_argument('--dataset_name', default='PathMNIST', type=str, help='name of dataset')
+parser.add_argument('--dataset_name', default='TissueMNIST', type=str, help='name of dataset')
 parser.add_argument('--data_seed', default=0, type=int, help='random seed data partitioning procedure')
 parser.add_argument('--training_seed', default=0, type=int, help='random seed for training procedure')
 
@@ -63,11 +66,14 @@ parser.add_argument('--train_dir',
 
 #data paths
 parser.add_argument('--l_train_dataset_path', default='', type=str)
+parser.add_argument('--u_train_dataset_path', default='', type=str)
 parser.add_argument('--val_dataset_path', default='', type=str)
 parser.add_argument('--test_dataset_path', default='', type=str)
 
 #shared config
 parser.add_argument('--labeledtrain_batchsize', default=50, type=int)
+parser.add_argument('--unlabeledtrain_batchsize', default=50, type=int)
+parser.add_argument("--em", default=0, type=float, help="coefficient of entropy minimization. If you try VAT + EM, set 0.06")
 
 #VAT config
 parser.add_argument('--lr', default=3e-4, type=float, help='learning rate')
@@ -80,6 +86,13 @@ parser.add_argument('--lr_cycle_epochs', default=10000, type=int) #following Mix
 
 parser.add_argument('--wd', default=5e-4, type=float, help='weight decay')
 parser.add_argument('--optimizer_type', default='SGD', choices=['SGD', 'Adam'], type=str) 
+
+
+parser.add_argument('--lambda_u_max', default=1, type=float, help='coefficient of unlabeled loss')
+
+parser.add_argument('--unlabeledloss_warmup_schedule_type', default='NoWarmup', choices=['NoWarmup', 'Linear', 'Sigmoid', ], type=str) 
+
+parser.add_argument('--unlabeledloss_warmup_pos', default=0.4, type=float, help='position at which unlabeled loss warmup ends') #following MixMatch and FixMatch repo
 
 
 
@@ -222,23 +235,32 @@ def main(args):
         transforms.Normalize(mean=dataset_mean, std=dataset_std)
     ])
     
-  
+    
+    class TransformTwice:
+        def __init__(self, transform_fn):
+            self.transform_fn = transform_fn
+        
+        def __call__(self, x):
+            out1 = self.transform_fn(x)
+            out2 = self.transform_fn(x)
+        
+            return out1, out2
 
     l_train_dataset = dataset(args.dataset_name, args.l_train_dataset_path, transform_fn=transform_labeledtrain)
+    u_train_dataset = dataset(args.dataset_name, args.u_train_dataset_path, transform_fn=TransformTwice(transform_labeledtrain))
     val_dataset = dataset(args.dataset_name, args.val_dataset_path, transform_fn=transform_eval)
     test_dataset = dataset(args.dataset_name, args.test_dataset_path, transform_fn=transform_eval)
     
     #########################################for loop from here##################################
-    
     global_best_val_raw_acc_list_parallel = []
     global_best_test_raw_acc_at_val_list_parallel = []
 
-    
+  
     global_best_val_raw_acc = 0
     global_best_test_raw_acc_at_val = 0
 #     global_best_train_raw_acc_at_val = 0
 
-    
+
     #create global writer
     global_writer = SummaryWriter(args.train_dir)
 
@@ -256,10 +278,11 @@ def main(args):
     while total_used_time <= args.total_hour * 3600: #run for 100 hour
         lr = sample_loguniform(low=-5, high=-2, size=1, coefficient=3, base=10)
         wd = sample_loguniform(low=-6, high=-3, size=1, coefficient=4, base=10)
+        lambda_u_max = sample_loguniform(low=-1, high=1, size=1, coefficient=8, base=10)
         
-        print('!!!!!!!!!!This run using lr: {}, wd: {}!!!!!!!!!!'.format(lr, wd))
+        print('!!!!!!!!!!This run using lr: {}, wd: {}, lambda_u_max: {}!!!!!!!!!!'.format(lr, wd, lambda_u_max))
         
-        hypercombo_iteratethrough_list.append({'lr':lr, 'wd':wd})#newly added
+        hypercombo_iteratethrough_list.append({'lr':lr, 'wd':wd, 'lambda_u_max':lambda_u_max})#newly added
         save_pickle(os.path.join(args.train_dir, 'global_stats'), 'hypercombo_iteratethrough_list.pkl', hypercombo_iteratethrough_list)
         this_hypercombo_starttime = time.time()
         
@@ -267,23 +290,29 @@ def main(args):
         this_hypercombo_best_test_raw_acc_at_val_list_parallel = []
 
         
+        ssl_obj = MT()
+
         args.lr = lr
         args.wd = wd
+        args.lambda_u_max = lambda_u_max
 
-        experiment_name = "lr-{}_wd-{}".format(args.lr, args.wd)
+        experiment_name = "LambdaUMax-{}_lr-{}_wd-{}_em-{}".format(args.lambda_u_max, args.lr, args.wd, args.em)
 
         args.experiment_dir = os.path.join(args.train_dir, 'hypercombos', experiment_name)
 
         #brief summary:
         brief_summary = {}
         brief_summary['dataset_name'] = args.dataset_name
-        brief_summary['algorithm'] = 'Supervised'
+        brief_summary['algorithm'] = 'MeanTeacher'
         brief_summary['hyperparameters'] = {
             'optimizer': args.optimizer_type,
             'lr_schedule_type': args.lr_schedule_type,
             'lr_cycle_epochs': args.lr_cycle_epochs,
+            'unlabeledloss_warmup_schedule_type':args.unlabeledloss_warmup_schedule_type,
+            'unlabeledloss_warmup_pos': args.unlabeledloss_warmup_pos,
             'lr': args.lr,
             'wd': args.wd,
+            'lambda_u_max': args.lambda_u_max,
 
         }
 
@@ -298,12 +327,13 @@ def main(args):
         args.writer = SummaryWriter(args.experiment_dir)
 
         print('Created dataset')
-        print("labeled data : {}, training data : {}".format(
-        len(l_train_dataset), len(l_train_dataset)))
+        print("labeled data : {}, unlabeled data : {}, training data : {}".format(
+        len(l_train_dataset), len(u_train_dataset), len(l_train_dataset)+len(u_train_dataset)))
         print("validation data : {}, test data : {}".format(len(val_dataset), len(test_dataset)))
     
 
         l_loader = DataLoader(l_train_dataset, args.labeledtrain_batchsize, shuffle=True, drop_last=True, num_workers=args.num_workers)
+        u_loader = DataLoader(u_train_dataset, args.unlabeledtrain_batchsize, shuffle=True, drop_last=True, num_workers=args.num_workers)
         val_loader = DataLoader(val_dataset, 128, shuffle=False, drop_last=False, num_workers=args.num_workers)
         test_loader = DataLoader(test_dataset, 128, shuffle=False, drop_last=False, num_workers=args.num_workers)
 
@@ -344,7 +374,7 @@ def main(args):
 
 
         #instantiate the ema model object
-#         ema_model = ModelEMA(args, model, args.ema_decay)
+        ema_model = ModelEMA(args, model, args.ema_decay)
 
         args.start_epoch = 0
 
@@ -354,7 +384,6 @@ def main(args):
 
         this_hypercombo_best_val_raw_acc_list_parallel.append(best_val_raw_acc)
         this_hypercombo_best_test_raw_acc_at_val_list_parallel.append(best_test_raw_acc_at_val)
-
         
         current_count=0 #for early stopping, when continue training
 
@@ -366,6 +395,7 @@ def main(args):
                 checkpoint = torch.load(args.resume_checkpoint_fullpath)
                 args.start_epoch = checkpoint['epoch']
                 model.load_state_dict(checkpoint['state_dict'])
+                ema_model.ema.load_state_dict(checkpoint['ema_state_dict'])
 
                 best_val_raw_acc = checkpoint['best_val_raw_acc']
                 best_test_raw_acc_at_val = checkpoint['best_test_raw_acc_at_val']
@@ -381,7 +411,7 @@ def main(args):
         logger.info("***** Running training *****")
         logger.info(f"  Task = {args.dataset_name}")
         logger.info(f"  Num Epochs = {args.train_epoch}")
-        logger.info(f"  Batch size per GPU (labeled) = {args.labeledtrain_batchsize}")
+        logger.info(f"  Batch size per GPU (labeled+unlabeled) = {args.labeledtrain_batchsize + args.unlabeledtrain_batchsize}")
 
 
         is_best = False
@@ -390,9 +420,9 @@ def main(args):
 
 
         for epoch in range(args.start_epoch, args.train_epoch):
-            
+        
             #train
-            train_labeled_loss_list = train_one_epoch(args, weights, l_loader, model, optimizer, scheduler, epoch)
+            train_total_loss_list, train_labeled_loss_list, train_unlabeled_loss_unscaled_list, train_unlabeled_loss_scaled_list = train_one_epoch(args, weights, ssl_obj, l_loader, u_loader, model, ema_model, optimizer, scheduler, epoch)
 
             
             #val
@@ -401,13 +431,13 @@ def main(args):
             #test
             test_loss, test_raw_acc, test_true_labels, test_raw_predictions = eval_model(args, test_loader, model, epoch, evaluation_criterion='balanced_accuracy')
 
+
             if val_raw_acc > best_val_raw_acc:
                 is_best = True
 
                 best_val_raw_acc = val_raw_acc
                 best_test_raw_acc_at_val = test_raw_acc
 #                 best_train_raw_acc_at_val = train_raw_acc
-
 
 
 
@@ -420,9 +450,13 @@ def main(args):
 
             start_time = time.time() #reinitialize
 
+
             logger.info('RAW Best , validation/test/train %.2f %.2f' % (best_val_raw_acc, best_test_raw_acc_at_val))
 
+            args.writer.add_scalar('train/3.total_loss', np.mean(train_total_loss_list), epoch)
             args.writer.add_scalar('train/4.labeled_loss', np.mean(train_labeled_loss_list), epoch)
+            args.writer.add_scalar('train/5.unlabeled_loss_unscaled', np.mean(train_unlabeled_loss_unscaled_list), epoch)
+            args.writer.add_scalar('train/6.unlabele_loss_scaled', np.mean(train_unlabeled_loss_scaled_list), epoch)
 
             args.writer.add_scalar('val/1.val_raw_acc', val_raw_acc, epoch)
             args.writer.add_scalar('val/3.val_loss', val_loss, epoch)
@@ -430,10 +464,9 @@ def main(args):
             args.writer.add_scalar('test/3.test_loss', test_loss, epoch)
 
             brief_summary["number_of_data"] = {
-        "labeled":len(l_train_dataset), 
+        "labeled":len(l_train_dataset), "unlabeled":len(u_train_dataset),
         "validation":len(val_dataset), "test":len(test_dataset)
     }
-
             brief_summary['best_test_raw_acc_at_val'] = best_test_raw_acc_at_val
             brief_summary['best_val_raw_acc'] = best_val_raw_acc
 
@@ -448,6 +481,7 @@ def main(args):
                 {
                 'epoch': epoch+1,
                 'state_dict': model.state_dict(),
+                'ema_state_dict': ema_model.ema.state_dict(),
                 'best_val_raw_acc': best_val_raw_acc,
                 'best_test_raw_acc_at_val': best_test_raw_acc_at_val,
                 'current_count':current_count,
@@ -473,7 +507,7 @@ def main(args):
 
         
         brief_summary["number_of_data"] = {
-        "labeled":len(l_train_dataset), 
+        "labeled":len(l_train_dataset), "unlabeled":len(u_train_dataset),
         "validation":len(val_dataset), "test":len(test_dataset)
     }
         
@@ -484,7 +518,7 @@ def main(args):
         global_best_val_raw_acc_list_parallel.append(this_hypercombo_best_val_raw_acc_list_parallel)
         global_best_test_raw_acc_at_val_list_parallel.append(this_hypercombo_best_test_raw_acc_at_val_list_parallel)
 
-        
+
         save_pickle(os.path.join(args.train_dir, 'global_stats'), 'global_best_val_raw_acc_list_parallel.pkl', global_best_val_raw_acc_list_parallel)
         save_pickle(os.path.join(args.train_dir, 'global_stats'), 'global_best_test_raw_acc_at_val_list_parallel.pkl', global_best_test_raw_acc_at_val_list_parallel)
         
@@ -493,11 +527,10 @@ def main(args):
 
         with open(os.path.join(args.experiment_dir + "brief_summary.json"), "w") as f:
             json.dump(brief_summary, f)
-
             
     save_pickle(os.path.join(args.train_dir, 'global_stats'), 'total_time.pkl', [total_used_time])
     
-    
+
                     
 if __name__ == '__main__':
     args = parser.parse_args()
@@ -505,7 +538,6 @@ if __name__ == '__main__':
     args.use_pretrained = str2bool(args.use_pretrained)
     args.nimg_per_epoch = dataset_configs[args.dataset_name]['nimg_per_epoch'] #total size of labeled + unlabeled set for TissueMNIST
     args.num_classes = dataset_configs[args.dataset_name]['num_classes']
-    
     
     cuda = torch.cuda.is_available()
     
@@ -526,9 +558,15 @@ if __name__ == '__main__':
     logger.info(dict(args._get_kwargs()))
 
     
-    
     if args.training_seed is not None:
         print('setting training seed{}'.format(args.training_seed), flush=True)
         set_seed(args.training_seed)
         
     main(args)
+    
+    
+    
+    
+
+
+
