+++ src/ssl_bench/MixMatch/MixMatch.py
@@ -28,12 +28,11 @@ from torchvision import transforms
 
 from torch.utils.tensorboard import SummaryWriter
 
-from src.dataset import data as dataset
-from src.randaugment import RandAugmentMC
-from src.FixMatch.libml.utils import save_pickle
-from src.FixMatch.libml.utils import train_one_epoch, eval_model
-from src.FixMatch.libml.utils import EarlyStopping
-from src.config import dataset_configs
+from ssl_bench.dataset import data as dataset
+from ssl_bench.MixMatch.libml.utils import save_pickle
+from ssl_bench.MixMatch.libml.utils import train_one_epoch, eval_model
+from ssl_bench.MixMatch.libml.utils import EarlyStopping
+from ssl_bench.config import dataset_configs
 
 
 logger = logging.getLogger(__name__)
@@ -72,7 +71,8 @@ parser.add_argument('--labeledtrain_batchsize', default=50, type=int)
 parser.add_argument('--unlabeledtrain_batchsize', default=50, type=int)
 parser.add_argument("--em", default=0, type=float, help="coefficient of entropy minimization. If you try VAT + EM, set 0.06")
 
-#FM config
+#PL config
+
 parser.add_argument('--lr', default=3e-4, type=float, help='learning rate')
 parser.add_argument('--lr_warmup_epochs', default=0, type=float,
                     help='warmup epoch for learning rate schedule') #following MixMatch and FixMatch repo
@@ -80,18 +80,14 @@ parser.add_argument('--lr_warmup_epochs', default=0, type=float,
 parser.add_argument('--lr_schedule_type', default='CosineLR', choices=['CosineLR', 'FixedLR'], type=str) 
 parser.add_argument('--lr_cycle_epochs', default=10000, type=int) #following MixMatch and FixMatch repo
 
+
 parser.add_argument('--wd', default=5e-4, type=float, help='weight decay')
 parser.add_argument('--optimizer_type', default='SGD', choices=['SGD', 'Adam'], type=str) 
 
 
 parser.add_argument('--temperature', default=0.95, type=float, help='temperature for label guessing')
 
-parser.add_argument('--mu', default=7, type=int,
-                    help='coefficient of unlabeled batch size')
-
-parser.add_argument('--threshold', default=0.95, type=float,
-                    help='pseudo label threshold')
-
+parser.add_argument('--alpha', default=0.75, type=float)
 
 
 parser.add_argument('--lambda_u_max', default=1, type=float, help='coefficient of unlabeled loss')
@@ -132,6 +128,8 @@ def sample_uniform(low=0.0, high=1.0, size=1, decimal=1):
     
     return round(np.random.uniform(low=low, high=high, size=size)[0],1)
 
+
+
 def str2bool(s):
     if s == 'True':
         return True
@@ -142,11 +140,11 @@ def str2bool(s):
     
 
 #checked
-def save_checkpoint(state, is_best, checkpoint, filename='last_checkpoint.pth.tar'):
-    filepath = os.path.join(checkpoint, filename)
+def save_checkpoint(state, is_best, checkpoint_dir, filename='last_checkpoint.pth.tar'):
+    filepath = os.path.join(checkpoint_dir, filename)
     torch.save(state, filepath)
     if is_best:
-        shutil.copyfile(filepath, os.path.join(checkpoint,
+        shutil.copyfile(filepath, os.path.join(checkpoint_dir,
                                                'model_best.pth.tar'))
         
 #checked
@@ -154,7 +152,6 @@ def set_seed(seed):
     random.seed(seed)
     np.random.seed(seed)
     torch.manual_seed(seed)
-    
     
     
 #learning rate schedule   
@@ -188,7 +185,7 @@ def get_fixed_lr(optimizer,
 def create_model(args):
     if args.arch == 'resnet18':
         from torchvision import models
-
+        
         model = models.resnet18(pretrained=args.use_pretrained)
         #https://stackoverflow.com/questions/52548174/how-to-remove-the-last-fc-layer-from-a-resnet-model-in-pytorch
         model.fc = torch.nn.Linear(512, args.num_classes)
@@ -203,10 +200,10 @@ def create_model(args):
                                         dropout=0.0,
                                         num_classes=args.num_classes)
         
-        
     else:
         raise NameError('Note implemented yet')
     
+
     
     logger.info("Total params: {:.2f}M".format(
         sum(p.numel() for p in model.parameters())/1e6))
@@ -226,7 +223,7 @@ def main(args):
     dataset_mean = dataset_configs[args.dataset_name]['dataset_mean']
     dataset_std = dataset_configs[args.dataset_name]['dataset_std']
     image_size = dataset_configs[args.dataset_name]['image_size']
-
+    
     
     transform_labeledtrain = transforms.Compose([
         transforms.RandomHorizontalFlip(),
@@ -243,65 +240,59 @@ def main(args):
     ])
     
     
-    class TransformFixMatch(object):
-        def __init__(self, mean, std):
-            self.weak = transforms.Compose([
-                transforms.RandomHorizontalFlip(),
-                transforms.RandomCrop(size=image_size,
-                                      padding=int(image_size*0.125),
-                                      padding_mode='reflect')])
-            self.strong = transforms.Compose([
-                transforms.RandomHorizontalFlip(),
-                transforms.RandomCrop(size=image_size,
-                                      padding=int(image_size*0.125),
-                                      padding_mode='reflect'),
-                RandAugmentMC(n=2, m=10)])
-            self.normalize = transforms.Compose([
-                transforms.ToTensor(),
-                transforms.Normalize(mean=mean, std=std)])
-
+    class TransformTwice:
+        def __init__(self, transform_fn):
+            self.transform_fn = transform_fn
+        
         def __call__(self, x):
-            weak = self.weak(x)
-            strong = self.strong(x)
-            return self.normalize(weak), self.normalize(strong)
-
-      
+            out1 = self.transform_fn(x)
+            out2 = self.transform_fn(x)
+        
+            return out1, out2
+        
+    
+    
     l_train_dataset = dataset(args.dataset_name, args.l_train_dataset_path, transform_fn=transform_labeledtrain)
-    u_train_dataset = dataset(args.dataset_name, args.u_train_dataset_path, transform_fn=TransformFixMatch(mean=dataset_mean, std=dataset_std))
+    u_train_dataset = dataset(args.dataset_name, args.u_train_dataset_path, transform_fn=TransformTwice(transform_labeledtrain))
     val_dataset = dataset(args.dataset_name, args.val_dataset_path, transform_fn=transform_eval)
     test_dataset = dataset(args.dataset_name, args.test_dataset_path, transform_fn=transform_eval)
-        
-        
-    #########################################for loop from here##################################  
+    
+    #########################################for loop from here##################################   
+    
     global_best_val_raw_acc_list_parallel = []
     global_best_test_raw_acc_at_val_list_parallel = []
 
     
     global_best_val_raw_acc = 0
     global_best_test_raw_acc_at_val = 0
+#     global_best_train_raw_acc_at_val = 0
+
 
     #create global writer
     global_writer = SummaryWriter(args.train_dir)
+    
 
     #start timing
     hypercombo_iteratethrough_list = [] #newly added
     hypercombo_iteratethrough_time_list = []
-   
+    
     start_time = time.time()
     total_used_time = 0
     
     #Lr: loguniform base 10, [3*10^(-5) to 3*10^(-2)]
     #Wd: loguniform base 10, [4*10^(-6) to 4*10^(-3)]
-    #Lambda_u_max: loguniform base 10, [1*10^(-1) to 1*10^(1)]
+    #Lambda_u_max: loguniform base 10, [7.5*10^(0)  to 7.5*10^(2)]
+    #Alpha: uniform [0.1, 1], round to 1 decimal point
     
     while total_used_time <= args.total_hour * 3600: #run for 100 hour
-        
         lr = sample_loguniform(low=-5, high=-2, size=1, coefficient=3, base=10)
         wd = sample_loguniform(low=-6, high=-3, size=1, coefficient=4, base=10)
-        lambda_u_max = sample_loguniform(low=-1, high=1, size=1, coefficient=1, base=10)
+        lambda_u_max = sample_loguniform(low=0, high=2, size=1, coefficient=7.5, base=10)
+        alpha = sample_uniform(low=0.1, high=1.0, size=1, decimal=1)
         
-        print('!!!!!!!!!!This run using lr: {}, wd: {}, lambda_u_max: {}!!!!!!!!!!'.format(lr, wd, lambda_u_max))
-        hypercombo_iteratethrough_list.append({'lr':lr, 'wd':wd, 'lambda_u_max':lambda_u_max})#newly added
+        print('!!!!!!!!!!This run using lr: {}, wd: {}, lambda_u_max: {}, alpha: {}!!!!!!!!!!'.format(lr, wd, lambda_u_max, alpha))
+        
+        hypercombo_iteratethrough_list.append({'lr':lr, 'wd':wd, 'lambda_u_max':lambda_u_max, 'alpha':alpha})#newly added
         save_pickle(os.path.join(args.train_dir, 'global_stats'), 'hypercombo_iteratethrough_list.pkl', hypercombo_iteratethrough_list)
         this_hypercombo_starttime = time.time()
         
@@ -312,27 +303,28 @@ def main(args):
         args.lr = lr
         args.wd = wd
         args.lambda_u_max = lambda_u_max
+        args.alpha = alpha
 
-        experiment_name = "LambdaUMax-{}_lr-{}_wd-{}_temperature-{}_mu-{}_threshold-{}_em-{}".format(args.lambda_u_max, args.lr, args.wd,  args.temperature, args.mu, args.threshold, args.em)
+        experiment_name = "LambdaUMax-{}_lr-{}_wd-{}_alpha-{}_temperature-{}_em-{}".format(args.lambda_u_max, args.lr, args.wd, args.alpha, args.temperature, args.em)
 
         args.experiment_dir = os.path.join(args.train_dir, 'hypercombos', experiment_name)
 
         #brief summary:
         brief_summary = {}
         brief_summary['dataset_name'] = args.dataset_name
-        brief_summary['algorithm'] = 'FixMatch'
+        brief_summary['algorithm'] = 'MixMatch'
         brief_summary['hyperparameters'] = {
             'optimizer': args.optimizer_type,
             'lr_schedule_type': args.lr_schedule_type,
             'lr_cycle_epochs': args.lr_cycle_epochs,
             'unlabeledloss_warmup_schedule_type':args.unlabeledloss_warmup_schedule_type,
             'unlabeledloss_warmup_pos': args.unlabeledloss_warmup_pos,
-            'lambda_u_max': args.lambda_u_max,        
+            'lambda_u_max': args.lambda_u_max,
             'lr': args.lr,
             'wd': args.wd,
             'temperature': args.temperature,
-            'mu': args.mu,
-            'threshold':args.threshold,
+            'alpha': args.alpha,
+
         }
 
 
@@ -342,6 +334,7 @@ def main(args):
         else:
             args.resume_checkpoint_fullpath = None
 
+
         os.makedirs(args.experiment_dir, exist_ok=True)
         args.writer = SummaryWriter(args.experiment_dir)
 
@@ -349,8 +342,8 @@ def main(args):
         print("labeled data : {}, unlabeled data : {}, training data : {}".format(
         len(l_train_dataset), len(u_train_dataset), len(l_train_dataset)+len(u_train_dataset)))
         print("validation data : {}, test data : {}".format(len(val_dataset), len(test_dataset)))
-
-
+    
+    
         l_loader = DataLoader(l_train_dataset, args.labeledtrain_batchsize, shuffle=True, drop_last=True, num_workers=args.num_workers)
         u_loader = DataLoader(u_train_dataset, args.unlabeledtrain_batchsize, shuffle=True, drop_last=True, num_workers=args.num_workers)
         val_loader = DataLoader(val_dataset, 128, shuffle=False, drop_last=False, num_workers=args.num_workers)
@@ -378,8 +371,6 @@ def main(args):
         else:
             raise NameError('Not supported optimizer setting')
 
-
-
         #lr_schedule_type choice
         if args.lr_schedule_type == 'CosineLR':
             scheduler = get_cosine_schedule_with_warmup(optimizer, args.lr_warmup_epochs, args.lr_cycle_epochs)
@@ -396,6 +387,7 @@ def main(args):
 
         args.start_epoch = 0
 
+        
         best_val_raw_acc = 0
         best_test_raw_acc_at_val = 0
 #                 best_train_raw_acc_at_val = 0
@@ -403,7 +395,8 @@ def main(args):
         this_hypercombo_best_val_raw_acc_list_parallel.append(best_val_raw_acc)
         this_hypercombo_best_test_raw_acc_at_val_list_parallel.append(best_test_raw_acc_at_val)
 
-        current_count=0 #for early stopping, when continue training   
+        
+        current_count=0 #for early stopping, when continue training
         
         if args.resume_checkpoint_fullpath is not None:
             try:
@@ -413,8 +406,6 @@ def main(args):
                 args.start_epoch = checkpoint['epoch']
                 model.load_state_dict(checkpoint['state_dict'])
 #                 ema_model.ema.load_state_dict(checkpoint['ema_state_dict'])
-
-#                         best_train_ema_acc_at_val = checkpoint['best_train_ema_acc_at_val']
 
                 best_val_raw_acc = checkpoint['best_val_raw_acc']
                 best_test_raw_acc_at_val = checkpoint['best_test_raw_acc_at_val']
@@ -443,33 +434,33 @@ def main(args):
             #train
             train_total_loss_list, train_labeled_loss_list, train_unlabeled_loss_unscaled_list, train_unlabeled_loss_scaled_list = train_one_epoch(args, weights, l_loader, u_loader, model, optimizer, scheduler, epoch)
 
-
+        
             #val
             val_loss, val_raw_acc, val_true_labels, val_raw_predictions = eval_model(args, val_loader, model, epoch, evaluation_criterion='balanced_accuracy')
-            
+
+
             #test
             test_loss, test_raw_acc, test_true_labels, test_raw_predictions = eval_model(args, test_loader, model, epoch, evaluation_criterion='balanced_accuracy')
 
 
             if val_raw_acc > best_val_raw_acc:
-                is_best = True
-                
+                is_best=True
+
                 best_val_raw_acc = val_raw_acc
                 best_test_raw_acc_at_val = test_raw_acc
-#                 best_train_raw_acc_at_val = train_raw_acc
 
-    
+
+
             this_hypercombo_best_val_raw_acc_list_parallel.append(best_val_raw_acc)
             this_hypercombo_best_test_raw_acc_at_val_list_parallel.append(best_test_raw_acc_at_val)
 
-                     
             elapsed_time = time.time() - start_time
             total_used_time += elapsed_time
 
             start_time = time.time() #reinitialize
 
-            logger.info('RAW Best , validation/test %.2f %.2f' % (best_val_raw_acc, best_test_raw_acc_at_val))
 
+            logger.info('RAW Best , validation/test %.2f %.2f' % (best_val_raw_acc, best_test_raw_acc_at_val))
 
             args.writer.add_scalar('train/3.total_loss', np.mean(train_total_loss_list), epoch)
             args.writer.add_scalar('train/4.labeled_loss', np.mean(train_labeled_loss_list), epoch)
@@ -512,10 +503,9 @@ def main(args):
             #early stopping
             if early_stopping.early_stop:
                 break
-                
+
             if total_used_time > args.total_hour * 3600:
                 break
-                
 
 
         this_hypercombo_endtime = time.time()
@@ -527,7 +517,6 @@ def main(args):
         "labeled":len(l_train_dataset), "unlabeled":len(u_train_dataset),
         "validation":len(val_dataset), "test":len(test_dataset)
     }
-
         brief_summary['best_test_raw_acc_at_val'] = best_test_raw_acc_at_val
         brief_summary['best_val_raw_acc'] = best_val_raw_acc
         
@@ -535,25 +524,24 @@ def main(args):
         global_best_val_raw_acc_list_parallel.append(this_hypercombo_best_val_raw_acc_list_parallel)
         global_best_test_raw_acc_at_val_list_parallel.append(this_hypercombo_best_test_raw_acc_at_val_list_parallel)
 
-        
         save_pickle(os.path.join(args.train_dir, 'global_stats'), 'global_best_val_raw_acc_list_parallel.pkl', global_best_val_raw_acc_list_parallel)
         save_pickle(os.path.join(args.train_dir, 'global_stats'), 'global_best_test_raw_acc_at_val_list_parallel.pkl', global_best_test_raw_acc_at_val_list_parallel)
-
+    
+        
         args.writer.close()
 
         with open(os.path.join(args.experiment_dir + "brief_summary.json"), "w") as f:
             json.dump(brief_summary, f)
-
             
     save_pickle(os.path.join(args.train_dir, 'global_stats'), 'total_time.pkl', [total_used_time])
     
-    
+
 if __name__ == '__main__':
     args = parser.parse_args()
     
     args.use_pretrained = str2bool(args.use_pretrained)
     args.nimg_per_epoch = dataset_configs[args.dataset_name]['nimg_per_epoch'] #total size of labeled + unlabeled set for TissueMNIST
-    args.num_classes = dataset_configs[args.dataset_name]['num_classes']
+    args.num_classes = dataset_configs[args.dataset_name]['num_classes']    
     
     cuda = torch.cuda.is_available()
     
@@ -574,9 +562,11 @@ if __name__ == '__main__':
     logger.info(dict(args._get_kwargs()))
 
     
+    
     if args.training_seed is not None:
         print('setting training seed{}'.format(args.training_seed), flush=True)
         set_seed(args.training_seed)
+    
 
     main(args)
