+++ src/ssl_bench/apply_clahe.py
