+++ src/ssl_bench/archive/dataset.py
