+++ src/ssl_bench/utils/apply_clahe.py
